Error handling done, most of the program done and working. Awaiting further tests and polish before finalizing.
