+++ CMP1903M A01 2223/CMP1903M A01 2223/CMP1903M A01 2223.csproj
@@ -45,7 +45,7 @@
   <ItemGroup>
     <Compile Include="Card.cs" />
     <Compile Include="Pack.cs" />
-    <Compile Include="Testing.cs" />
+    <Compile Include="Test.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
